test: split e2e sync tests by feature AAS-16
